* Точность подгона скрола * Заполнение директории с помощью FSProtocol * Добавление в FSProtocol *dir() и generic флагов для open и opendir * auto -> type & * Чистка устаревшего дебага * FS_ATTR* -> FSProtocol::FSEntryFlags::* * Убран сброс позиции при навигации * Выделение в списке файлов с аттрибутами Hidden и Readonly * Фикс выделения всего списка * Полностью убрана зависимость от sigc++ в пользу легковестного signal_slot

+++ trunk/gcc/Sources/LampFM/LocalFSProtocol.cpp
@@ -1,34 +1,40 @@
 #include "LocalFSProtocol.h"
 #include <stdio.h>
+#include <stdlib.h>
 #include <unistd.h>
 #include <fcntl.h>
+#include <fs.h>
 
 
 
+LocalFSProtocol::LocalFSProtocol()
+{
+
+}
 
 
-void *LocalFSProtocol::open(const char *f, int mode)
+void *LocalFSProtocol::open(const char *f, FSProtocol::OpenMode mode)
 {
     char m[4] = {0};
     int at = 0;
 
-    if(mode & FSProtocol::Mode::READ)
+    if(mode & FSProtocol::OpenMode::Read)
         m[at++] = 'r';
 
-    if(mode & FSProtocol::Mode::WRITE)
+    if(mode & FSProtocol::OpenMode::Write)
         m[at++] = 'w';
 
-    if(mode & FSProtocol::Mode::CREATE)
+    if(mode & FSProtocol::OpenMode::Create)
         m[at++] = '+';
 
-    if(mode & FSProtocol::Mode::TRUNC) {
+    if(mode & FSProtocol::OpenMode::Trunc) {
         m[0] = 'w';
         m[1] = '+';
         m[3] = 0;
     }
 
 
-    FILE *fp = fopen(f, m);
+    ::FILE *fp = fopen(f, m);
     return fp;
 }
 
@@ -41,11 +47,85 @@ int LocalFSProtocol::unlink(const char *f)
 
 int LocalFSProtocol::mkdir(const char *d, bool recursive)
 {
-    return ::mkdir(d, 0777);
+    if(!recursive)
+        return ::mkdir(d, 0777);
+
+    return -1;
 }
 
 
 int LocalFSProtocol::rmdir(const char *d, bool recursive)
 {
-    return ::rmdir(d);
+    if(!recursive)
+        return ::rmdir(d);
+
+    return -1;
+}
+
+
+void *LocalFSProtocol::opendir(const char *dir, const char *mask)
+{
+    ((void)mask);
+
+    int d = fs_opendir(dir);
+
+    printf("LocalFSProtocol::opendir: %d\n", d);
+
+    if(d > 0) {
+        void **r = (void **)malloc(sizeof(void*));
+        *r = (void*)d;
+        return (void*)r;
+    }
+
+    return 0;
+}
+
+
+int LocalFSProtocol::readdir(void *hdir, FSProtocol::FSEntry *entry)
+{
+    if(!hdir)
+        return -1;
+
+    int handle = *(int*)hdir;
+
+    FS_INFO info;
+    int r = fs_readdir(handle, &info);
+
+    if(!r) {
+        entry->flags = FSProtocol::FSEntryFlags::File;
+        entry->name  = info.name;
+        entry->size  = info.size;
+
+        if(info.attr & FS_ATTR_FOLDER)
+            entry->flags |= FSProtocol::FSEntryFlags::Dir;
+
+        if(info.attr & FS_ATTR_HIDE)
+            entry->flags |= FSProtocol::FSEntryFlags::Hidden;
+
+        if(info.attr & FS_ATTR_READONLY) {
+            entry->flags |= FSProtocol::FSEntryFlags::Readonly;
+            printf("READONLY FUCK \n");
+        }
+
+        /*info.attr &= ~FS_ATTR_FOLDER;
+        info.attr &= ~FS_ATTR_HIDE;
+        info.attr &= ~FS_ATTR_READONLY;
+
+        printf("info.attr(%s): %d\n", info.name, info.attr);*/
+
+    }
+
+    return r;
+}
+
+
+int LocalFSProtocol::closedir(void *hdir)
+{
+    if(!hdir)
+        return -1;
+
+    int handle = *(int*)hdir;
+    free(hdir);
+
+    return fs_closedir(handle);
 }
